requirements.txt
selenium
pyautogui
webdriver_manager
PyYAML
validate_email
openai>=1.0.0
pypdf>=3.0.0